refactored and debugged functionality of the app
