+++ src/main/java/org/store/structure/IntroApplication.java
@@ -1,4 +1,4 @@
-package org.store.structure.config;
+package org.store.structure;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
