+++ src/main/java/org/store/structure/config/OpenApiConfig.java
@@ -1,9 +1,9 @@
 package org.store.structure.config;
 
-import java.util.List;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
 import io.swagger.v3.oas.models.servers.Server;
+import java.util.List;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
