+++ src/main/resources/application.properties
@@ -3,10 +3,7 @@ spring.datasource.username=root
 spring.datasource.password=Piltrafilla1]
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
-
 spring.jpa.hibernate.ddl-auto=validate
 spring.jpa.show-sql=true
 
 spring.jackson.deserialization.fail-on-unknown-properties=true
-
-springdoc.swagger-ui.displayOperationId=true
